DAVE-36: Apply background scaling and bug fixing
