+++ src/main/resources/static/scripts/config.js
@@ -16,5 +16,6 @@ CONFIG = {
   ENERGY_FILTER_STEP: 0.005,
   BULK_ANALYSIS_ENABLED: false,
   LOG_TO_SERVER_ENABLED: true, //If true, python server logs will be visible from GUI Log tab
-  LOG_LEVEL: -1 // PYTHON SERVER LOG LEVEL -> ALL = -1, DEBUG = 0, INFO = 1, WARN = 2, ERROR = 3, NONE = 4
+  LOG_LEVEL: -1, // PYTHON SERVER LOG LEVEL -> ALL = -1, DEBUG = 0, INFO = 1, WARN = 2, ERROR = 3, NONE = 4
+  DENY_BCK_IF_SUBS: true //Avoid set background file if lightcurve bck data is allready substracted
 }
