+++ src/main/resources/templates/includes/output_panel.html
@@ -10,7 +10,8 @@
 <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/html2canvas.js') }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/canvas2image.js') }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/jspdf_1_3_3.min.js') }}"></script>
-<script type="text/javascript" src="{{ url_for('static', filename='scripts/outputpanel.js') }}?v={{ get_version() }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='scripts/outputPanels/outputPanel.js') }}?v={{ get_version() }}"></script>
+<script type="text/javascript" src="{{ url_for('static', filename='scripts/outputPanels/wfOutputPanel.js') }}?v={{ get_version() }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='scripts/plots/plotter.js') }}?v={{ get_version() }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='scripts/plots/plot.js') }}?v={{ get_version() }}"></script>
 <script type="text/javascript" src="{{ url_for('static', filename='scripts/plots/plotWithSettings.js') }}?v={{ get_version() }}"></script>
