+++ src/main/resources/templates/master_page.html
@@ -19,9 +19,9 @@
     <script type=text/javascript src="{{ url_for('static', filename='scripts/external/jquery-3.1.1.min.js') }}"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/jquery-ui-1.12.1.min.js') }}"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/bootstrap-3.3.7.min.js') }}"></script>
-    <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/logger-1.15.min.js') }}"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/async-2.5.0.min.js') }}"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/waiting_dialog.js') }}?v={{ get_version() }}"></script>
+    <script type="text/javascript" src="{{ url_for('static', filename='scripts/external/logger.js') }}?v={{ get_version() }}"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='scripts/config.js') }}?v={{ get_version() }}"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='scripts/service.js') }}?v={{ get_version() }}"></script>
     <script type="text/javascript" src="{{ url_for('static', filename='scripts/schema.js') }}?v={{ get_version() }}"></script>
