+++ src/test/resources/python_examples/testing_fits.py
@@ -1,8 +1,8 @@
 from astropy.io import fits
-hdulist = fits.open('std1_ao9_01_01.lc')
-tbdata = hdulist[1].data
-print(tbdata.field(0))
-print(tbdata.field(1))
-print(tbdata.field(2))
-print(tbdata.field(3))
+import numpy as np
 
+fits_file = '../std1_ao9_01_09.lc'
+hdulist = fits.open(fits_file)
+tbdata = hdulist['RATE'].data
+tbdata['RATE'] = 20
+hdulist.writeto('../src20.lc')
